Move Validator to its own file
